separated socket types into different files and changed imports in the files where they were used
