+++ client/src/pages/Game.tsx
@@ -2,12 +2,12 @@ import { useLocation } from "react-router-dom"
 import useSocket from "../hooks/socket/useSocket";
 import Board from "../components/game/Board/Board";
 import { useEffect, useState } from "react";
-import { GAME_ACCEPTED } from "../constants/actions";
-import { GameAcceptedData } from "../types/api/socket";
+import { GAME_ACCEPTED, GET_MOVES } from "../constants/actions";
 import { CellHandle } from "../types/game/component-types";
 import { useAppDispatch, useAppSelector } from "../app/hooks";
 import { selectColor, selectMatrix, setInitialGameState } from "../features/game/gameSlice";
 import { WHITE } from "../constants/game";
+import { GameAcceptedData } from "../types/socket/receiveData";
 
 export default function Game(){
     const location = useLocation()
