+++ client/src/types/hooks/socket.ts
@@ -1,4 +1,4 @@
-import { Action } from "../api/socket";
+import { Action } from "../socket/global"
 
 export type Listeners = Record<Action, Listener<unknown>>
 export type Listener<D> = (data: D)=>void
