+++ client/src/types/socket/global.ts
@@ -1,0 +1,7 @@
+import { GAME_ACCEPTED, GAME_FOUND, GET_MOVES } from "../../constants/actions"
+
+export type Action = typeof GAME_ACCEPTED | typeof GAME_FOUND | typeof GET_MOVES
+
+export interface SocketDataBase {
+    action : Action
+}
