+++ client/src/types/socket/receiveData.ts
@@ -1,13 +1,6 @@
-import { GAME_ACCEPTED, GAME_FOUND } from "../../constants/actions"
 import { BLACK, WHITE } from "../../constants/game"
 import { Matrix } from "../game/game"
-
-export type SocketData = {
-    action : Action,
-    data : Data
-}
-
-export type Action = typeof GAME_ACCEPTED | typeof GAME_FOUND
+import { SocketDataBase } from "./global"
 
 export type GameFoundData = {
     game_id : number
@@ -18,4 +11,8 @@ export type GameAcceptedData = {
     color : typeof BLACK | typeof WHITE
 }
 
-export type Data = GameFoundData
+export type Data = GameFoundData | GameAcceptedData
+
+export interface SocketData extends SocketDataBase {
+    data : Data
+}
